feat: Se mejora funcionalidad de búsqueda de productos, se mejora custom hook useProducts y servicio searchProducts
